requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==5.2.13
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.1
gunicorn==25.3.0
idna==3.11
packaging==26.0
requests==2.33.1
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
